requirements.txt
unidecode
numpy
sentence-transformers
feedparser
faiss-cpu
newspaper3k
lxml[html_clean]
